move pagination footers to their own files
