+++ src/components/atoms/CopyableRow/copyableRow.spec.ts
@@ -1,0 +1,22 @@
+describe('Copyable Row', () => {
+  describe('When Copy is Enabled', () => {
+    it('copies content onClick', () => {});
+    it('shows copy icon', () => {});
+  });
+
+  describe('When Copy is Disabled', () => {
+    it('does not show copy icon', () => {});
+  });
+});
+
+// Goals:
+/**
+ * - figure out how to do the shared layout without remounting
+ * - finish account creation
+ * - toggle notification when transfer is processed
+ * - invalidate query and refetch users transactions
+ * - add custom content to "notification" to support this
+ * - add unit tests (~3-5)
+ * - finish pagination pess
+ * - merge!
+ */
