+++ src/components/atoms/PaginatedFooter/paginatedFooter.spec.ts
@@ -1,0 +1,16 @@
+// TODO: add first unit test
+
+// make sure its disabled when it should be
+// its
+
+describe('PaginatedFooter', () => {
+  it('disables the left arrow on page 1', () => {});
+
+  it('enables the left arrow on page > 1', () => {});
+
+  it('disables the right arrow on page 10', () => {});
+
+  it('enables the right arrow on page > 10', () => {});
+
+  it('calls proper callbacks on presss', () => {});
+});
