+++ src/components/index.ts
@@ -13,6 +13,7 @@ import {
   FeatureFlagsEmptyState,
   FeatureFlagToggleList,
   RecentUsersEmptyState,
+  RecentUsersPagination,
   RecentUsersRow,
   RecentUsersTable,
   ReferralsEmptyState,
@@ -22,6 +23,7 @@ import {
   SelectedUsersReferralsTable,
   SelectedUsersTransfersTable,
   SelectedUserTransferRow,
+  TransactionsPagination,
   TransfersEmptyState,
   UserFlagsCard,
   UserProfileSummary,
@@ -53,6 +55,7 @@ export {
   PaginatedFooter,
   Popup,
   RecentUsersEmptyState,
+  RecentUsersPagination,
   RecentUsersRow,
   RecentUsersSection,
   RecentUsersTable,
@@ -68,6 +71,7 @@ export {
   SelectedUserTransfersSection,
   SingleColumnContentWrapper,
   Spinner,
+  TransactionsPagination,
   TransfersEmptyState,
   TwoColumnContentWrapper,
   UserFlagsCard,
