+++ src/components/molecules/RecentUsersPagination/index.ts
@@ -1,0 +1,3 @@
+import RecentUsersPagination from './RecentUsersPagination';
+
+export default RecentUsersPagination;
