+++ src/components/molecules/TransactionPagination/index.ts
@@ -1,0 +1,3 @@
+import TransactionsPagination from './TransactionPagination';
+
+export default TransactionsPagination;
